requirements.txt
streamlit
tensorflow==2.12.0
opencv-python-headless
numpy==1.23.5
pillow